Split up long battery tracing plugin.
