+++ ui/src/core/default_plugins.ts
@@ -30,6 +30,9 @@ export const defaultPlugins = [
   'com.android.AndroidPerf',
   'com.android.AndroidPerfTraceCounters',
   'com.android.AndroidStartup',
+  'com.android.Bluetooth',
+  'com.android.ContainedTraces',
+  'com.android.DayExplorer',
   'com.android.GpuWorkPeriod',
   'com.android.LargeScreensPerf',
   'com.android.PinAndroidPerfMetrics',
