+++ ui/src/plugins/com.android.AndroidLongBatterySupport/OWNERS
@@ -1,0 +1,2 @@
+simonmacm@google.com
+rzuklie@google.com
