+++ ui/src/plugins/com.android.AndroidLongBatteryTracing/index.ts
@@ -12,33 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {uuidv4} from '../../base/uuid';
 import {Trace} from '../../public/trace';
 import StandardGroupsPlugin from '../dev.perfetto.StandardGroups';
 import {PerfettoPlugin} from '../../public/plugin';
-import {Engine} from '../../trace_processor/engine';
-import {createQuerySliceTrack} from '../../components/tracks/query_slice_track';
-import {CounterOptions} from '../../components/tracks/base_counter_track';
-import {createQueryCounterTrack} from '../../components/tracks/query_counter_track';
-import {TrackNode} from '../../public/workspace';
-import {STR, LONG} from '../../trace_processor/query_result';
-import {AreaSelection, areaSelectionsEqual} from '../../public/selection';
-import {Flamegraph} from '../../widgets/flamegraph';
-import {
-  metricsFromTableOrSubquery,
-  QueryFlamegraph,
-} from '../../components/query_flamegraph';
-
-const DAY_EXPLORER_TRACK_KIND = 'day_explorer_counter_track';
-
-interface ContainedTrace {
-  uuid: string;
-  subscription: string;
-  trigger: string;
-  // NB: these are millis.
-  ts: number;
-  dur: number;
-}
+import {STR} from '../../trace_processor/query_result';
+import SupportPlugin from '../com.android.AndroidLongBatterySupport';
 
 const PACKAGE_LOOKUP = `
   create or replace perfetto table package_name_lookup as
@@ -175,134 +153,6 @@ const NETWORK_SUMMARY = `
       group by 1, 2, 3
   )
   select * from final where ts is not null`;
-
-const MODEM_RIL_STRENGTH = `
-  DROP VIEW IF EXISTS ScreenOn;
-  CREATE VIEW ScreenOn AS
-  SELECT ts, dur FROM (
-      SELECT
-          ts, value,
-          LEAD(ts, 1, TRACE_END()) OVER (ORDER BY ts)-ts AS dur
-      FROM counter, track ON (counter.track_id = track.id)
-      WHERE track.name = 'ScreenState'
-  ) WHERE value = 2;
-
-  DROP VIEW IF EXISTS RilSignalStrength;
-  CREATE VIEW RilSignalStrength AS
-  With RilMessages AS (
-      SELECT
-          ts, slice.name,
-          LEAD(ts, 1, TRACE_END()) OVER (ORDER BY ts)-ts AS dur
-      FROM slice, track
-      ON (slice.track_id = track.id)
-      WHERE track.name = 'RIL'
-        AND slice.name GLOB 'UNSOL_SIGNAL_STRENGTH*'
-  ),
-  BandTypes(band_ril, band_name) AS (
-      VALUES ("CellSignalStrengthLte:", "LTE"),
-              ("CellSignalStrengthNr:", "NR")
-  ),
-  ValueTypes(value_ril, value_name) AS (
-      VALUES ("rsrp=", "rsrp"),
-              ("rssi=", "rssi")
-  ),
-  Extracted AS (
-      SELECT ts, dur, band_name, value_name, (
-          SELECT CAST(SUBSTR(key_str, start_idx+1, end_idx-start_idx-1) AS INT64) AS value
-          FROM (
-              SELECT key_str, INSTR(key_str, "=") AS start_idx, INSTR(key_str, " ") AS end_idx
-              FROM (
-                  SELECT SUBSTR(band_str, INSTR(band_str, value_ril)) AS key_str
-                  FROM (SELECT SUBSTR(name, INSTR(name, band_ril)) AS band_str)
-              )
-          )
-      ) AS value
-      FROM RilMessages
-      JOIN BandTypes
-      JOIN ValueTypes
-  )
-  SELECT
-  ts, dur, band_name, value_name, value,
-  value_name || "=" || IIF(value = 2147483647, "unknown", ""||value) AS name,
-  ROW_NUMBER() OVER (ORDER BY ts) as id,
-  DENSE_RANK() OVER (ORDER BY band_name, value_name) AS track_id
-  FROM Extracted;
-
-  DROP TABLE IF EXISTS RilScreenOn;
-  CREATE VIRTUAL TABLE RilScreenOn
-  USING SPAN_JOIN(RilSignalStrength PARTITIONED track_id, ScreenOn)`;
-
-const MODEM_RIL_CHANNELS_PREAMBLE = `
-  CREATE OR REPLACE PERFETTO FUNCTION EXTRACT_KEY_VALUE(source STRING, key_name STRING) RETURNS STRING AS
-  SELECT SUBSTR(trimmed, INSTR(trimmed, "=")+1, INSTR(trimmed, ",") - INSTR(trimmed, "=") - 1)
-  FROM (SELECT SUBSTR($source, INSTR($source, $key_name)) AS trimmed);`;
-
-const MODEM_RIL_CHANNELS = `
-  With RawChannelConfig AS (
-      SELECT ts, slice.name AS raw_config
-      FROM slice, track
-      ON (slice.track_id = track.id)
-      WHERE track.name = 'RIL'
-      AND slice.name LIKE 'UNSOL_PHYSICAL_CHANNEL_CONFIG%'
-  ),
-  Attributes(attribute, attrib_name) AS (
-      VALUES ("mCellBandwidthDownlinkKhz", "downlink"),
-          ("mCellBandwidthUplinkKhz", "uplink"),
-          ("mNetworkType", "network"),
-          ("mBand", "band")
-  ),
-  Slots(idx, slot_name) AS (
-      VALUES (0, "primary"),
-          (1, "secondary 1"),
-          (2, "secondary 2")
-  ),
-  Stage1 AS (
-      SELECT *, IFNULL(EXTRACT_KEY_VALUE(STR_SPLIT(raw_config, "}, {", idx), attribute), "") AS name
-      FROM RawChannelConfig
-      JOIN Attributes
-      JOIN Slots
-  ),
-  Stage2 AS (
-      SELECT *, LAG(name) OVER (PARTITION BY idx, attribute ORDER BY ts) AS last_name
-      FROM Stage1
-  ),
-  Stage3 AS (
-      SELECT *, LEAD(ts, 1, TRACE_END()) OVER (PARTITION BY idx, attribute ORDER BY ts) - ts AS dur
-      FROM Stage2 WHERE name != last_name
-  )
-  SELECT ts, dur, slot_name || "-" || attrib_name || "=" || name AS name
-  FROM Stage3`;
-
-const MODEM_CELL_RESELECTION = `
-  with base as (
-    select
-        ts,
-        s.name as raw_ril,
-        ifnull(str_split(str_split(s.name, 'CellIdentityLte{', 1), ', operatorNames', 0),
-            str_split(str_split(s.name, 'CellIdentityNr{', 1), ', operatorNames', 0)) as cell_id
-    from track t join slice s on t.id = s.track_id
-    where t.name = 'RIL' and s.name like '%DATA_REGISTRATION_STATE%'
-  ),
-  base2 as (
-    select
-        ts,
-        raw_ril,
-        case
-            when cell_id like '%earfcn%' then 'LTE ' || cell_id
-            when cell_id like '%nrarfcn%' then 'NR ' || cell_id
-            when cell_id is null then 'Unknown'
-            else cell_id
-        end as cell_id
-    from base
-  ),
-  base3 as (
-    select ts, cell_id , lag(cell_id) over (order by ts) as lag_cell_id, raw_ril
-    from base2
-  )
-  select ts, 0 as dur, cell_id as name, raw_ril
-  from base3
-  where cell_id != lag_cell_id
-  order by ts`;
 
 const SUSPEND_RESUME = `
   SELECT
@@ -482,460 +332,17 @@ const WAKEUPS_COLUMNS = [
   'backoff_millis',
 ];
 
-function bleScanQuery(condition: string) {
-  return `
-  with step1 as (
-      select
-          ts,
-          extract_arg(arg_set_id, 'ble_scan_state_changed.attribution_node[0].tag') as name,
-          extract_arg(arg_set_id, 'ble_scan_state_changed.is_opportunistic') as opportunistic,
-          extract_arg(arg_set_id, 'ble_scan_state_changed.is_filtered') as filtered,
-          extract_arg(arg_set_id, 'ble_scan_state_changed.state') as state
-      from track t join slice s on t.id = s.track_id
-      where t.name = 'Statsd Atoms'
-      and s.name = 'ble_scan_state_changed'
-  ),
-  step2 as (
-      select
-          ts,
-          name,
-          state,
-          opportunistic,
-          filtered,
-          lead(ts) over (partition by name order by ts) - ts as dur
-      from step1
-  )
-  select ts, dur, name from step2 where state = 'ON' and ${condition} and dur is not null`;
-}
-
-const BLE_RESULTS = `
-  with step1 as (
-      select
-          ts,
-          extract_arg(arg_set_id, 'ble_scan_result_received.attribution_node[0].tag') as name,
-          extract_arg(arg_set_id, 'ble_scan_result_received.num_results') as num_results
-      from track t join slice s on t.id = s.track_id
-      where t.name = 'Statsd Atoms'
-      and s.name = 'ble_scan_result_received'
-  )
-  select
-      ts,
-      0 as dur,
-      name || ' (' || num_results || ' results)' as name
-  from step1`;
-
-const BT_A2DP_AUDIO = `
-  with step1 as (
-    select
-        ts,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_a2dp_playback_state_changed.playback_state') as playback_state,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_a2dp_playback_state_changed.audio_coding_mode') as audio_coding_mode,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_a2dp_playback_state_changed.metric_id') as metric_id
-    from track t join slice s on t.id = s.track_id
-    where t.name = 'Statsd Atoms'
-    and s.name = 'bluetooth_a2dp_playback_state_changed'
-  ),
-  step2 as (
-    select
-        ts,
-        lead(ts) over (partition by metric_id order by ts) - ts as dur,
-        playback_state,
-        audio_coding_mode,
-        metric_id
-    from step1
-  )
-  select
-    ts,
-    dur,
-    audio_coding_mode as name
-  from step2
-  where playback_state = 'PLAYBACK_STATE_PLAYING'`;
-
-const BT_CONNS_ACL = `
-    with acl1 as (
-        select
-            ts,
-            EXTRACT_ARG(arg_set_id, 'bluetooth_acl_connection_state_changed.state') as state,
-            EXTRACT_ARG(arg_set_id, 'bluetooth_acl_connection_state_changed.transport') as transport,
-            EXTRACT_ARG(arg_set_id, 'bluetooth_acl_connection_state_changed.metric_id') as metric_id
-        from track t join slice s on t.id = s.track_id
-        where t.name = 'Statsd Atoms'
-        and s.name = 'bluetooth_acl_connection_state_changed'
-    ),
-    acl2 as (
-        select
-            ts,
-            lead(ts) over (partition by metric_id, transport order by ts) - ts as dur,
-            state,
-            transport,
-            metric_id
-        from acl1
-    )
-    select
-        ts,
-        dur,
-        'Device ' || metric_id ||
-          ' (' || case transport when 'TRANSPORT_TYPE_BREDR' then 'Classic' when 'TRANSPORT_TYPE_LE' then 'BLE' end || ')' as name
-    from acl2
-    where state != 'CONNECTION_STATE_DISCONNECTED' and dur is not null`;
-
-const BT_CONNS_SCO = `
-  with sco1 as (
-    select
-        ts,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_sco_connection_state_changed.state') as state,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_sco_connection_state_changed.codec') as codec,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_sco_connection_state_changed.metric_id') as metric_id
-    from track t join slice s on t.id = s.track_id
-    where t.name = 'Statsd Atoms'
-    and s.name = 'bluetooth_sco_connection_state_changed'
-  ),
-  sco2 as (
-    select
-        ts,
-        lead(ts) over (partition by metric_id, codec order by ts) - ts as dur,
-        state,
-        codec,
-        metric_id
-    from sco1
-  )
-  select
-    ts,
-    dur,
-    case state when 'CONNECTION_STATE_CONNECTED' then '' when 'CONNECTION_STATE_CONNECTING' then 'Connecting ' when 'CONNECTION_STATE_DISCONNECTING' then 'Disconnecting ' else 'unknown ' end ||
-      'Device ' || metric_id || ' (' ||
-      case codec when 'SCO_CODEC_CVSD' then 'CVSD' when 'SCO_CODEC_MSBC' then 'MSBC' end || ')' as name
-  from sco2
-  where state != 'CONNECTION_STATE_DISCONNECTED' and dur is not null`;
-
-const BT_LINK_LEVEL_EVENTS = `
-  with base as (
-    select
-        ts,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_link_layer_connection_event.direction') as direction,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_link_layer_connection_event.type') as type,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_link_layer_connection_event.hci_cmd') as hci_cmd,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_link_layer_connection_event.hci_event') as hci_event,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_link_layer_connection_event.hci_ble_event') as hci_ble_event,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_link_layer_connection_event.cmd_status') as cmd_status,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_link_layer_connection_event.reason_code') as reason_code,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_link_layer_connection_event.metric_id') as metric_id
-    from track t join slice s on t.id = s.track_id
-    where t.name = 'Statsd Atoms'
-    and s.name = 'bluetooth_link_layer_connection_event'
-  )
-  select
-    *,
-    0 as dur,
-    'Device '|| metric_id as name
-  from base`;
-
-const BT_LINK_LEVEL_EVENTS_COLUMNS = [
-  'direction',
-  'type',
-  'hci_cmd',
-  'hci_event',
-  'hci_ble_event',
-  'cmd_status',
-  'reason_code',
-  'metric_id',
-];
-
-const BT_QUALITY_REPORTS = `
-  with base as (
-      select
-          ts,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.quality_report_id') as quality_report_id,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.packet_types') as packet_types,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.connection_handle') as connection_handle,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.connection_role') as connection_role,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.tx_power_level') as tx_power_level,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.rssi') as rssi,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.snr') as snr,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.unused_afh_channel_count') as unused_afh_channel_count,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.afh_select_unideal_channel_count') as afh_select_unideal_channel_count,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.lsto') as lsto,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.connection_piconet_clock') as connection_piconet_clock,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.retransmission_count') as retransmission_count,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.no_rx_count') as no_rx_count,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.nak_count') as nak_count,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.flow_off_count') as flow_off_count,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.buffer_overflow_bytes') as buffer_overflow_bytes,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_quality_report_reported.buffer_underflow_bytes') as buffer_underflow_bytes
-      from track t join slice s on t.id = s.track_id
-      where t.name = 'Statsd Atoms'
-      and s.name = 'bluetooth_quality_report_reported'
-  )
-  select
-      *,
-      0 as dur,
-      'Connection '|| connection_handle as name
-  from base`;
-
-const BT_QUALITY_REPORTS_COLUMNS = [
-  'quality_report_id',
-  'packet_types',
-  'connection_handle',
-  'connection_role',
-  'tx_power_level',
-  'rssi',
-  'snr',
-  'unused_afh_channel_count',
-  'afh_select_unideal_channel_count',
-  'lsto',
-  'connection_piconet_clock',
-  'retransmission_count',
-  'no_rx_count',
-  'nak_count',
-  'flow_off_count',
-  'buffer_overflow_bytes',
-  'buffer_underflow_bytes',
-];
-
-const BT_RSSI_REPORTS = `
-  with base as (
-    select
-        ts,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_device_rssi_reported.connection_handle') as connection_handle,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_device_rssi_reported.hci_status') as hci_status,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_device_rssi_reported.rssi') as rssi,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_device_rssi_reported.metric_id') as metric_id
-    from track t join slice s on t.id = s.track_id
-    where t.name = 'Statsd Atoms'
-    and s.name = 'bluetooth_device_rssi_reported'
-  )
-  select
-    *,
-    0 as dur,
-    'Connection '|| connection_handle as name
-  from base`;
-
-const BT_RSSI_REPORTS_COLUMNS = [
-  'connection_handle',
-  'hci_status',
-  'rssi',
-  'metric_id',
-];
-
-const BT_CODE_PATH_COUNTER = `
-  with base as (
-    select
-        ts,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_code_path_counter.key') as key,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_code_path_counter.number') as number
-    from track t join slice s on t.id = s.track_id
-    where t.name = 'Statsd Atoms'
-    and s.name = 'bluetooth_code_path_counter'
-  )
-  select
-    *,
-    0 as dur,
-    key as name
-  from base`;
-
-const BT_CODE_PATH_COUNTER_COLUMNS = ['key', 'number'];
-
-const BT_HAL_CRASHES = `
-  with base as (
-      select
-          ts,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_hal_crash_reason_reported.metric_id') as metric_id,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_hal_crash_reason_reported.error_code') as error_code,
-          EXTRACT_ARG(arg_set_id, 'bluetooth_hal_crash_reason_reported.vendor_error_code') as vendor_error_code
-      from track t join slice s on t.id = s.track_id
-      where t.name = 'Statsd Atoms'
-      and s.name = 'bluetooth_hal_crash_reason_reported'
-  )
-  select
-      *,
-      0 as dur,
-      'Device ' || metric_id as name
-  from base`;
-
-const BT_HAL_CRASHES_COLUMNS = ['metric_id', 'error_code', 'vendor_error_code'];
-
-const BT_BYTES = `
-  with step1 as (
-    select
-        ts,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_bytes_transfer.uid') as uid,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_bytes_transfer.tx_bytes') as tx_bytes,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_bytes_transfer.rx_bytes') as rx_bytes
-    from track t join slice s on t.id = s.track_id
-    where t.name = 'Statsd Atoms'
-    and s.name = 'bluetooth_bytes_transfer'
-  ),
-  step2 as (
-    select
-        ts,
-        lead(ts) over (partition by uid order by ts) - ts as dur,
-        uid,
-        lead(tx_bytes) over (partition by uid order by ts) - tx_bytes as tx_bytes,
-        lead(rx_bytes) over (partition by uid order by ts) - rx_bytes as rx_bytes
-    from step1
-  ),
-  step3 as (
-    select
-        ts,
-        dur,
-        uid % 100000 as uid,
-        sum(tx_bytes) as tx_bytes,
-        sum(rx_bytes) as rx_bytes
-    from step2
-    where tx_bytes >=0 and rx_bytes >=0
-    group by 1,2,3
-    having tx_bytes > 0 or rx_bytes > 0
-  )
-    select
-        ts,
-        dur,
-        format("%s: TX %d bytes / RX %d bytes", package_name, tx_bytes, rx_bytes) as name
-    from add_package_name!(step3)
-`;
-
-// See go/bt_system_context_report for reference on the bit-twiddling.
-const BT_ACTIVITY = `
-  create perfetto table bt_activity as
-  with step1 as (
-    select
-        EXTRACT_ARG(arg_set_id, 'bluetooth_activity_info.timestamp_millis') * 1000000 as ts,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_activity_info.bluetooth_stack_state') as bluetooth_stack_state,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_activity_info.controller_idle_time_millis') * 1000000 as controller_idle_dur,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_activity_info.controller_tx_time_millis') * 1000000 as controller_tx_dur,
-        EXTRACT_ARG(arg_set_id, 'bluetooth_activity_info.controller_rx_time_millis') * 1000000 as controller_rx_dur
-    from track t join slice s on t.id = s.track_id
-    where t.name = 'Statsd Atoms'
-    and s.name = 'bluetooth_activity_info'
-  ),
-  step2 as (
-    select
-        ts,
-        lead(ts) over (order by ts) - ts as dur,
-        bluetooth_stack_state,
-        lead(controller_idle_dur) over (order by ts) - controller_idle_dur as controller_idle_dur,
-        lead(controller_tx_dur) over (order by ts) - controller_tx_dur as controller_tx_dur,
-        lead(controller_rx_dur) over (order by ts) - controller_rx_dur as controller_rx_dur
-    from step1
-  )
-  select
-    ts,
-    dur,
-    bluetooth_stack_state & 0x0000000F as acl_active_count,
-    bluetooth_stack_state & 0x000000F0 >> 4 as acl_sniff_count,
-    bluetooth_stack_state & 0x00000F00 >> 8 as acl_ble_count,
-    bluetooth_stack_state & 0x0000F000 >> 12 as advertising_count,
-    case bluetooth_stack_state & 0x000F0000 >> 16
-      when 0 then 0
-      when 1 then 5
-      when 2 then 10
-      when 3 then 25
-      when 4 then 100
-      else -1
-    end as le_scan_duty_cycle,
-    bluetooth_stack_state & 0x00100000 >> 20 as inquiry_active,
-    bluetooth_stack_state & 0x00200000 >> 21 as sco_active,
-    bluetooth_stack_state & 0x00400000 >> 22 as a2dp_active,
-    bluetooth_stack_state & 0x00800000 >> 23 as le_audio_active,
-    max(0, 100.0 * controller_idle_dur / dur) as controller_idle_pct,
-    max(0, 100.0 * controller_tx_dur / dur) as controller_tx_pct,
-    max(0, 100.0 * controller_rx_dur / dur) as controller_rx_pct
-  from step2
-`;
-
 export default class implements PerfettoPlugin {
   static readonly id = 'com.android.AndroidLongBatteryTracing';
-  static readonly dependencies = [StandardGroupsPlugin];
+  static readonly dependencies = [StandardGroupsPlugin, SupportPlugin];
 
-  private readonly groups = new Map<string, TrackNode>();
-
-  private getOrCreateGroup(
-    ctx: Trace,
-    groupName: string,
-    groupCollapsed = true,
-  ): TrackNode {
-    const existingGroup = this.groups.get(groupName);
-    if (existingGroup) {
-      return existingGroup;
-    }
-
-    const group = new TrackNode({
-      name: groupName,
-      isSummary: true,
-      collapsed: groupCollapsed,
-    });
-    this.groups.set(groupName, group);
-    ctx.workspace.addChildInOrder(group);
-    return group;
-  }
-
-  private addTrack(
-    ctx: Trace,
-    track: TrackNode,
-    groupName?: string,
-    groupCollapsed = true,
-  ): void {
-    if (groupName) {
-      const group = this.getOrCreateGroup(ctx, groupName, groupCollapsed);
-      group.addChildInOrder(track);
-    } else {
-      ctx.workspace.addChildInOrder(track);
-    }
-  }
-
-  async addSliceTrack(
-    ctx: Trace,
-    name: string,
-    query: string,
-    groupName: string,
-    columns: string[] = [],
-    groupCollapsed = true,
-  ) {
-    const uri = `/long_battery_tracing_${name}`;
-    const track = await createQuerySliceTrack({
-      trace: ctx,
-      uri,
-      data: {
-        sqlSource: query,
-        columns: ['ts', 'dur', 'name', ...columns],
-      },
-      argColumns: columns,
-    });
-    ctx.tracks.registerTrack({
-      uri,
-      renderer: track,
-    });
-    const trackNode = new TrackNode({uri, name});
-    this.addTrack(ctx, trackNode, groupName, groupCollapsed);
-  }
-
-  async addCounterTrack(
-    ctx: Trace,
-    name: string,
-    query: string,
-    groupName: string,
-    options?: Partial<CounterOptions>,
-    groupCollapsed = true,
-  ) {
-    const uri = `/long_battery_tracing_${name}`;
-    const track = await createQueryCounterTrack({
-      trace: ctx,
-      uri,
-      data: {
-        sqlSource: query,
-        columns: ['ts', 'value'],
-      },
-      options,
-    });
-    ctx.tracks.registerTrack({
-      uri,
-      renderer: track,
-    });
-    const trackNode = new TrackNode({uri, name});
-    this.addTrack(ctx, trackNode, groupName, groupCollapsed);
+  private support(ctx: Trace) {
+    return ctx.plugins.getPlugin(SupportPlugin);
   }
 
   async addBatteryStatsState(
     ctx: Trace,
+    support: SupportPlugin,
     name: string,
     track: string,
     groupName: string,
@@ -944,7 +351,7 @@ export default class implements PerfettoPlugin {
     if (!features.has(`track.${track}`)) {
       return;
     }
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       name,
       `SELECT ts, safe_dur AS dur, value_name AS name
@@ -956,6 +363,7 @@ export default class implements PerfettoPlugin {
 
   async addBatteryStatsEvent(
     ctx: Trace,
+    support: SupportPlugin,
     name: string,
     track: string,
     groupName: string,
@@ -965,7 +373,7 @@ export default class implements PerfettoPlugin {
       return;
     }
 
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       name,
       `SELECT ts, safe_dur AS dur, str_value AS name
@@ -975,7 +383,11 @@ export default class implements PerfettoPlugin {
     );
   }
 
-  async addDeviceState(ctx: Trace, features: Set<string>): Promise<void> {
+  async addDeviceState(
+    ctx: Trace,
+    support: SupportPlugin,
+    features: Set<string>,
+  ): Promise<void> {
     if (!features.has('track.battery_stats.*')) {
       return;
     }
@@ -984,10 +396,10 @@ export default class implements PerfettoPlugin {
     const deviceStateGroup = ctx.plugins
       .getPlugin(StandardGroupsPlugin)
       .getOrCreateStandardGroup(ctx.workspace, 'DEVICE_STATE');
-    this.groups.set(groupName, deviceStateGroup);
+    support.groups.set(groupName, deviceStateGroup);
 
     const query = (name: string, track: string) =>
-      this.addBatteryStatsEvent(ctx, name, track, groupName, features);
+      this.addBatteryStatsEvent(ctx, support, name, track, groupName, features);
 
     const e = ctx.engine;
     await e.query(`INCLUDE PERFETTO MODULE android.battery_stats;`);
@@ -996,31 +408,31 @@ export default class implements PerfettoPlugin {
     await e.query(`INCLUDE PERFETTO MODULE android.battery.doze;`);
     await e.query(`INCLUDE PERFETTO MODULE android.screen_state;`);
 
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       'Screen state',
       `SELECT ts, dur, screen_state AS name FROM android_screen_state`,
       groupName,
     );
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       'Charging',
       `SELECT ts, dur, charging_state AS name FROM android_charging_states`,
       groupName,
     );
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       'Suspend / resume',
       SUSPEND_RESUME,
       groupName,
     );
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       'Doze light state',
       `SELECT ts, dur, light_idle_state AS name FROM android_light_idle_state`,
       groupName,
     );
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       'Doze deep state',
       `SELECT ts, dur, deep_idle_state AS name FROM android_deep_idle_state`,
@@ -1029,7 +441,7 @@ export default class implements PerfettoPlugin {
 
     query('Top app', 'battery_stats.top');
 
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       'Long wakelocks',
       `SELECT
@@ -1054,7 +466,7 @@ export default class implements PerfettoPlugin {
     ) {
       await e.query(`INCLUDE PERFETTO MODULE
          google3.wireless.android.telemetry.trace_extractor.modules.power.jobs;`);
-      await this.addSliceTrack(
+      await support.addSliceTrack(
         ctx,
         'Jobs',
         `SELECT ts, dur, tag AS name, uid FROM jobs`,
@@ -1066,7 +478,7 @@ export default class implements PerfettoPlugin {
     }
 
     if (features.has('atom.thermal_throttling_severity_state_changed')) {
-      await this.addSliceTrack(
+      await support.addSliceTrack(
         ctx,
         'Thermal throttling',
         THERMAL_THROTTLING,
@@ -1075,7 +487,7 @@ export default class implements PerfettoPlugin {
     }
   }
 
-  async addAtomCounters(ctx: Trace): Promise<void> {
+  async addAtomCounters(ctx: Trace, support: SupportPlugin): Promise<void> {
     const e = ctx.engine;
 
     await e.query(
@@ -1103,7 +515,7 @@ export default class implements PerfettoPlugin {
             ? {unit}
             : undefined;
 
-      await this.addCounterTrack(
+      await support.addCounterTrack(
         ctx,
         countersIt.ui_name,
         `select ts, ${unit === '%' ? 100.0 : 1.0} * counter_value as value
@@ -1115,7 +527,7 @@ export default class implements PerfettoPlugin {
     }
   }
 
-  async addAtomSlices(ctx: Trace): Promise<void> {
+  async addAtomSlices(ctx: Trace, support: SupportPlugin): Promise<void> {
     const e = ctx.engine;
 
     await e.query(
@@ -1170,7 +582,7 @@ export default class implements PerfettoPlugin {
                 as ${safeArg(arg)}`;
       }
 
-      await this.addSliceTrack(
+      await support.addSliceTrack(
         ctx,
         tracks.get(atom)!.ui_name,
         `select ts, dur, slice_name as name, ${args.map((a) => argSql(a)).join(', ')}
@@ -1183,195 +595,11 @@ export default class implements PerfettoPlugin {
     }
   }
 
-  async addDayExplorerCounters(
+  async addNetworkSummary(
     ctx: Trace,
-    groupName: string,
-    limit: number,
+    support: SupportPlugin,
+    features: Set<string>,
   ): Promise<void> {
-    await ctx.engine.query(
-      `INCLUDE PERFETTO MODULE
-          google3.wireless.android.telemetry.trace_extractor.modules.day_explorer.perfetto_ui_blames`,
-    );
-
-    const group = this.getOrCreateGroup(ctx, groupName);
-    await this.addDayExplorerRecursive(ctx, group, limit, -1n);
-  }
-
-  private async addDayExplorerRecursive(
-    ctx: Trace,
-    parent: TrackNode,
-    limit: number,
-    parentId: bigint,
-  ): Promise<void> {
-    const children = await ctx.engine.query(`
-      SELECT track_id, display_name, cast(round(total_energy_uws / 3600000) as int) as energy_mwh
-      FROM day_explorer_ui_hierarchy
-      WHERE (${parentId} >= 0 AND parent_id = ${parentId})
-         OR (${parentId} < 0 AND parent_id IS NULL)
-      ORDER BY energy_mwh DESC
-      LIMIT ${limit}
-    `);
-
-    const childIter = children.iter({
-      track_id: LONG,
-      display_name: STR,
-      energy_mwh: LONG,
-    });
-
-    for (; childIter.valid(); childIter.next()) {
-      const query = `
-        SELECT ts, power_mw AS value
-        FROM day_explorer_ui_hierarchy_per_ts
-        WHERE track_id = ${childIter.track_id}
-      `;
-      const groupKey = `_day_explorer_ui_hierarchy_under_${parentId}`;
-      const trackName = `${childIter.display_name} - ${childIter.energy_mwh}mWh`;
-      const node = await this.createDayExplorerTrack(
-        ctx,
-        trackName,
-        groupKey,
-        query,
-      );
-      parent.addChildInOrder(node);
-      await this.addDayExplorerRecursive(ctx, node, limit, childIter.track_id);
-    }
-  }
-
-  private async createDayExplorerTrack(
-    ctx: Trace,
-    name: string,
-    groupKey: string,
-    query: string,
-  ): Promise<TrackNode> {
-    const uri = `/day_explorer_${uuidv4()}`;
-    const renderer = await createQueryCounterTrack({
-      trace: ctx,
-      uri,
-      data: {
-        sqlSource: query,
-      },
-      columns: {
-        ts: 'ts',
-        value: 'value',
-      },
-      options: {
-        yRangeSharingKey: groupKey,
-      },
-    });
-
-    ctx.tracks.registerTrack({
-      uri,
-      renderer,
-      tags: {
-        kind: DAY_EXPLORER_TRACK_KIND,
-      },
-    });
-
-    return new TrackNode({
-      name,
-      uri,
-    });
-  }
-
-  private createDayExplorerFlameGraphPanel(trace: Trace) {
-    let previousSelection: AreaSelection | undefined;
-    let flamegraph: QueryFlamegraph | undefined;
-    return {
-      id: 'day_explorer_flamegraph_selection',
-      name: 'Day Explorer Flamegraph',
-      render: (selection: AreaSelection) => {
-        const selectionChanged =
-          previousSelection === undefined ||
-          !areaSelectionsEqual(previousSelection, selection);
-        previousSelection = selection;
-        if (selectionChanged) {
-          flamegraph = this.computeDayExplorerFlameGraph(trace, selection);
-        }
-
-        if (flamegraph === undefined) {
-          return undefined;
-        }
-
-        return {isLoading: false, content: flamegraph.render()};
-      },
-    };
-  }
-
-  computeDayExplorerFlameGraph(trace: Trace, currentSelection: AreaSelection) {
-    // The flame graph will be shown when any day explorer track is in the area
-    // selection. The selection is used to filter by time, but not by track. All
-    // day explorer tracks are considered for the graph.
-    let hasDayExplorer = false;
-    for (const trackInfo of currentSelection.tracks) {
-      if (trackInfo?.tags?.kind === DAY_EXPLORER_TRACK_KIND) {
-        hasDayExplorer = true;
-        break;
-      }
-    }
-    if (!hasDayExplorer) {
-      return undefined;
-    }
-    const metrics = metricsFromTableOrSubquery(
-      `
-        (
-          WITH
-            total_energy AS (
-              SELECT track_id, parent_id, display_name, SUM(energy_uws) AS energy_uws
-              FROM day_explorer_ui_hierarchy_per_ts
-              WHERE ts >= ${currentSelection.start}
-                AND ts <= ${currentSelection.end}
-              GROUP BY 1, 2, 3
-            ),
-            with_child AS (
-              SELECT
-                *,
-                (
-                  SELECT IFNULL(SUM(energy_uws), 0)
-                  FROM total_energy
-                  WHERE parent_id = P.track_id
-                ) AS child_energy
-              FROM total_energy AS P
-            )
-          SELECT
-            track_id AS id,
-            parent_id AS parentId,
-            display_name AS name,
-            cast(round((energy_uws - child_energy) / 1000) as int) AS self_count
-          FROM with_child
-        )
-      `,
-      [
-        {
-          name: 'Energy mWs',
-          unit: '',
-          columnName: 'self_count',
-        },
-      ],
-    );
-    return new QueryFlamegraph(trace, metrics, {
-      state: Flamegraph.createDefaultState(metrics),
-    });
-  }
-
-  async addDayExplorerBehaviors(ctx: Trace, groupName: string): Promise<void> {
-    const e = ctx.engine;
-
-    await e.query(
-      `INCLUDE PERFETTO MODULE
-          google3.wireless.android.telemetry.trace_extractor.modules.day_explorer.perfetto_ui_blames`,
-    );
-
-    await this.addSliceTrack(
-      ctx,
-      'Day Explorer Behaviors',
-      `select ts, dur, behavior as name from day_explorer_behaviors`,
-      groupName,
-      [],
-      false,
-    );
-  }
-
-  async addNetworkSummary(ctx: Trace, features: Set<string>): Promise<void> {
     if (!features.has('net.modem') && !features.has('net.wifi')) {
       return;
     }
@@ -1384,7 +612,7 @@ export default class implements PerfettoPlugin {
     await e.query(NETWORK_SUMMARY);
     await e.query(RADIO_TRANSPORT_TYPE);
 
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       'Default network',
       DEFAULT_NETWORK,
@@ -1392,10 +620,10 @@ export default class implements PerfettoPlugin {
     );
 
     if (features.has('atom.network_tethering_reported')) {
-      await this.addSliceTrack(ctx, 'Tethering', TETHERING, groupName);
+      await support.addSliceTrack(ctx, 'Tethering', TETHERING, groupName);
     }
     if (features.has('net.wifi')) {
-      await this.addCounterTrack(
+      await support.addCounterTrack(
         ctx,
         'Wifi total bytes',
         `select ts, sum(value) as value from network_summary where dev_type = 'wifi' group by 1`,
@@ -1407,7 +635,7 @@ export default class implements PerfettoPlugin {
       );
       const it = result.iter({pkg: STR});
       for (; it.valid(); it.next()) {
-        await this.addCounterTrack(
+        await support.addCounterTrack(
           ctx,
           `Top wifi: ${it.pkg}`,
           `select ts, value from network_summary where dev_type = 'wifi' and pkg = '${it.pkg}'`,
@@ -1418,6 +646,7 @@ export default class implements PerfettoPlugin {
     }
     this.addBatteryStatsState(
       ctx,
+      support,
       'Wifi interface',
       'battery_stats.wifi_radio',
       groupName,
@@ -1425,6 +654,7 @@ export default class implements PerfettoPlugin {
     );
     this.addBatteryStatsState(
       ctx,
+      support,
       'Wifi supplicant state',
       'battery_stats.wifi_suppl',
       groupName,
@@ -1432,13 +662,14 @@ export default class implements PerfettoPlugin {
     );
     this.addBatteryStatsState(
       ctx,
+      support,
       'Wifi strength',
       'battery_stats.wifi_signal_strength',
       groupName,
       features,
     );
     if (features.has('net.modem')) {
-      await this.addCounterTrack(
+      await support.addCounterTrack(
         ctx,
         'Modem total bytes',
         `select ts, sum(value) as value from network_summary where dev_type = 'modem' group by 1`,
@@ -1450,7 +681,7 @@ export default class implements PerfettoPlugin {
       );
       const it = result.iter({pkg: STR});
       for (; it.valid(); it.next()) {
-        await this.addCounterTrack(
+        await support.addCounterTrack(
           ctx,
           `Top modem: ${it.pkg}`,
           `select ts, value from network_summary where dev_type = 'modem' and pkg = '${it.pkg}'`,
@@ -1461,12 +692,13 @@ export default class implements PerfettoPlugin {
     }
     this.addBatteryStatsState(
       ctx,
+      support,
       'Cellular interface',
       'battery_stats.mobile_radio',
       groupName,
       features,
     );
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       'Cellular connection',
       `select ts, dur, name from radio_transport`,
@@ -1474,6 +706,7 @@ export default class implements PerfettoPlugin {
     );
     this.addBatteryStatsState(
       ctx,
+      support,
       'Cellular strength',
       'battery_stats.phone_signal_strength',
       groupName,
@@ -1481,48 +714,7 @@ export default class implements PerfettoPlugin {
     );
   }
 
-  async addModemRil(ctx: Trace, features: Set<string>): Promise<void> {
-    const groupName = 'Modem Detail';
-
-    if (!features.has('track.ril')) {
-      return;
-    }
-
-    const rilStrength = async (band: string, value: string) =>
-      await this.addSliceTrack(
-        ctx,
-        `Modem signal strength ${band} ${value}`,
-        `SELECT ts, dur, name FROM RilScreenOn WHERE band_name = '${band}' AND value_name = '${value}'`,
-        groupName,
-      );
-
-    const e = ctx.engine;
-
-    await e.query(MODEM_RIL_STRENGTH);
-    await e.query(MODEM_RIL_CHANNELS_PREAMBLE);
-
-    await rilStrength('LTE', 'rsrp');
-    await rilStrength('LTE', 'rssi');
-    await rilStrength('NR', 'rsrp');
-    await rilStrength('NR', 'rssi');
-
-    await this.addSliceTrack(
-      ctx,
-      'Modem channel config',
-      MODEM_RIL_CHANNELS,
-      groupName,
-    );
-
-    await this.addSliceTrack(
-      ctx,
-      'Modem cell reselection',
-      MODEM_CELL_RESELECTION,
-      groupName,
-      ['raw_ril'],
-    );
-  }
-
-  async addModemTeaData(ctx: Trace): Promise<void> {
+  async addModemTeaData(ctx: Trace, support: SupportPlugin): Promise<void> {
     const e = ctx.engine;
     const groupName = 'Modem Detail';
 
@@ -1536,7 +728,7 @@ export default class implements PerfettoPlugin {
     );
     const countersIt = counters.iter({name: STR});
     for (; countersIt.valid(); countersIt.next()) {
-      await this.addCounterTrack(
+      await support.addCounterTrack(
         ctx,
         countersIt.name,
         `select ts, value from pixel_modem_counters where name = '${countersIt.name}'`,
@@ -1548,7 +740,7 @@ export default class implements PerfettoPlugin {
     );
     const slicesIt = slices.iter({track_name: STR});
     for (; slicesIt.valid(); slicesIt.next()) {
-      await this.addSliceTrack(
+      await support.addSliceTrack(
         ctx,
         slicesIt.track_name,
         `select ts, dur, slice_name as name from pixel_modem_slices
@@ -1558,7 +750,7 @@ export default class implements PerfettoPlugin {
     }
   }
 
-  async addKernelWakelocks(ctx: Trace): Promise<void> {
+  async addKernelWakelocks(ctx: Trace, support: SupportPlugin): Promise<void> {
     const e = ctx.engine;
     await e.query(`INCLUDE PERFETTO MODULE android.kernel_wakelocks;`);
     const result = await e.query(
@@ -1566,7 +758,7 @@ export default class implements PerfettoPlugin {
     );
     const it = result.iter({name: STR, type: STR});
     for (; it.valid(); it.next()) {
-      await this.addCounterTrack(
+      await support.addCounterTrack(
         ctx,
         it.name,
         `SELECT ts, dur, held_ratio * 100 AS value
@@ -1580,6 +772,7 @@ export default class implements PerfettoPlugin {
 
   async addKernelWakelocksStatsd(
     ctx: Trace,
+    support: SupportPlugin,
     features: Set<string>,
   ): Promise<void> {
     if (!features.has('atom.kernel_wakelock')) {
@@ -1593,7 +786,7 @@ export default class implements PerfettoPlugin {
     const result = await e.query(KERNEL_WAKELOCKS_STATSD_SUMMARY);
     const it = result.iter({wakelock_name: STR});
     for (; it.valid(); it.next()) {
-      await this.addCounterTrack(
+      await support.addCounterTrack(
         ctx,
         `${it.wakelock_name} (statsd)`,
         `select ts, dur, value
@@ -1605,7 +798,11 @@ export default class implements PerfettoPlugin {
     }
   }
 
-  async addWakeups(ctx: Trace, features: Set<string>): Promise<void> {
+  async addWakeups(
+    ctx: Trace,
+    support: SupportPlugin,
+    features: Set<string>,
+  ): Promise<void> {
     if (!features.has('track.suspend_backoff')) {
       return;
     }
@@ -1642,7 +839,7 @@ export default class implements PerfettoPlugin {
     let labelOther = false;
     for (; it.valid(); it.next()) {
       labelOther = true;
-      await this.addSliceTrack(
+      await support.addSliceTrack(
         ctx,
         `Wakeup ${it.item}`,
         `${sqlPrefix} where item="${it.item}"`,
@@ -1651,7 +848,7 @@ export default class implements PerfettoPlugin {
       );
       items.push(it.item);
     }
-    await this.addSliceTrack(
+    await support.addSliceTrack(
       ctx,
       labelOther ? 'Other wakeups' : 'Wakeups',
       `${sqlPrefix} where item not in ('${items.join("','")}')`,
@@ -1660,7 +857,11 @@ export default class implements PerfettoPlugin {
     );
   }
 
-  async addHighCpu(ctx: Trace, features: Set<string>): Promise<void> {
+  async addHighCpu(
+    ctx: Trace,
+    support: SupportPlugin,
+    features: Set<string>,
+  ): Promise<void> {
     if (!features.has('atom.cpu_cycles_per_uid_cluster')) {
       return;
     }
@@ -1674,7 +875,7 @@ export default class implements PerfettoPlugin {
     );
     const it = result.iter({pkg: STR, cluster: STR});
     for (; it.valid(); it.next()) {
-      await this.addCounterTrack(
+      await support.addCounterTrack(
         ctx,
         `CPU (${it.cluster}): ${it.pkg}`,
         `select ts, value from high_cpu where pkg = "${it.pkg}" and cluster="${it.cluster}"`,
@@ -1684,281 +885,22 @@ export default class implements PerfettoPlugin {
     }
   }
 
-  async addBluetooth(ctx: Trace, features: Set<string>): Promise<void> {
-    if (
-      !Array.from(features.values()).some(
-        (f) => f.startsWith('atom.bluetooth_') || f.startsWith('atom.ble_'),
-      )
-    ) {
-      return;
-    }
-    const groupName = 'Bluetooth';
-    await this.addSliceTrack(
-      ctx,
-      'BLE Scans (opportunistic)',
-      bleScanQuery('opportunistic'),
-      groupName,
-    );
-    await this.addSliceTrack(
-      ctx,
-      'BLE Scans (filtered)',
-      bleScanQuery('filtered'),
-      groupName,
-    );
-    await this.addSliceTrack(
-      ctx,
-      'BLE Scans (unfiltered)',
-      bleScanQuery('not filtered'),
-      groupName,
-    );
-    await this.addSliceTrack(ctx, 'BLE Scan Results', BLE_RESULTS, groupName);
-    await this.addSliceTrack(ctx, 'Connections (ACL)', BT_CONNS_ACL, groupName);
-    await this.addSliceTrack(ctx, 'Connections (SCO)', BT_CONNS_SCO, groupName);
-    await this.addSliceTrack(
-      ctx,
-      'Link-level Events',
-      BT_LINK_LEVEL_EVENTS,
-      groupName,
-      BT_LINK_LEVEL_EVENTS_COLUMNS,
-    );
-    await this.addSliceTrack(ctx, 'A2DP Audio', BT_A2DP_AUDIO, groupName);
-    await this.addSliceTrack(
-      ctx,
-      'Bytes Transferred (L2CAP/RFCOMM)',
-      BT_BYTES,
-      groupName,
-    );
-    await ctx.engine.query(BT_ACTIVITY);
-    await this.addCounterTrack(
-      ctx,
-      'ACL Classic Active Count',
-      'select ts, dur, acl_active_count as value from bt_activity',
-      groupName,
-    );
-    await this.addCounterTrack(
-      ctx,
-      'ACL Classic Sniff Count',
-      'select ts, dur, acl_sniff_count as value from bt_activity',
-      groupName,
-    );
-    await this.addCounterTrack(
-      ctx,
-      'ACL BLE Count',
-      'select ts, dur, acl_ble_count as value from bt_activity',
-      groupName,
-    );
-    await this.addCounterTrack(
-      ctx,
-      'Advertising Instance Count',
-      'select ts, dur, advertising_count as value from bt_activity',
-      groupName,
-    );
-    await this.addCounterTrack(
-      ctx,
-      'LE Scan Duty Cycle Maximum',
-      'select ts, dur, le_scan_duty_cycle as value from bt_activity',
-      groupName,
-      {unit: '%'},
-    );
-    await this.addSliceTrack(
-      ctx,
-      'Inquiry Active',
-      "select ts, dur, 'Active' as name from bt_activity where inquiry_active",
-      groupName,
-    );
-    await this.addSliceTrack(
-      ctx,
-      'SCO Active',
-      "select ts, dur, 'Active' as name from bt_activity where sco_active",
-      groupName,
-    );
-    await this.addSliceTrack(
-      ctx,
-      'A2DP Active',
-      "select ts, dur, 'Active' as name from bt_activity where a2dp_active",
-      groupName,
-    );
-    await this.addSliceTrack(
-      ctx,
-      'LE Audio Active',
-      "select ts, dur, 'Active' as name from bt_activity where le_audio_active",
-      groupName,
-    );
-    await this.addCounterTrack(
-      ctx,
-      'Controller Idle Time',
-      'select ts, dur, controller_idle_pct as value from bt_activity',
-      groupName,
-      {yRangeSharingKey: 'bt_controller_time', unit: '%'},
-    );
-    await this.addCounterTrack(
-      ctx,
-      'Controller TX Time',
-      'select ts, dur, controller_tx_pct as value from bt_activity',
-      groupName,
-      {yRangeSharingKey: 'bt_controller_time', unit: '%'},
-    );
-    await this.addCounterTrack(
-      ctx,
-      'Controller RX Time',
-      'select ts, dur, controller_rx_pct as value from bt_activity',
-      groupName,
-      {yRangeSharingKey: 'bt_controller_time', unit: '%'},
-    );
-    await this.addSliceTrack(
-      ctx,
-      'Quality reports',
-      BT_QUALITY_REPORTS,
-      groupName,
-      BT_QUALITY_REPORTS_COLUMNS,
-    );
-    await this.addSliceTrack(
-      ctx,
-      'RSSI Reports',
-      BT_RSSI_REPORTS,
-      groupName,
-      BT_RSSI_REPORTS_COLUMNS,
-    );
-    await this.addSliceTrack(
-      ctx,
-      'HAL Crashes',
-      BT_HAL_CRASHES,
-      groupName,
-      BT_HAL_CRASHES_COLUMNS,
-    );
-    await this.addSliceTrack(
-      ctx,
-      'Code Path Counter',
-      BT_CODE_PATH_COUNTER,
-      groupName,
-      BT_CODE_PATH_COUNTER_COLUMNS,
-    );
-  }
-
-  async addContainedTraces(
-    ctx: Trace,
-    containedTraces: ContainedTrace[],
-  ): Promise<void> {
-    const bySubscription = new Map<string, ContainedTrace[]>();
-    for (const trace of containedTraces) {
-      if (!bySubscription.has(trace.subscription)) {
-        bySubscription.set(trace.subscription, []);
-      }
-      bySubscription.get(trace.subscription)!.push(trace);
-    }
-
-    for (const [subscription, traces] of bySubscription) {
-      await this.addSliceTrack(
-        ctx,
-        subscription,
-        traces
-          .map(
-            (t) => `SELECT
-          CAST(${t.ts} * 1e6 AS int) AS ts,
-          CAST(${t.dur} * 1e6 AS int) AS dur,
-          '${t.trigger === '' ? 'Trace' : t.trigger}' AS name,
-          'http://go/trace-uuid/${t.uuid}' AS link
-        `,
-          )
-          .join(' UNION ALL '),
-        'Other traces',
-        ['link'],
-      );
-    }
-  }
-
-  async findFeatures(e: Engine): Promise<Set<string>> {
-    const features = new Set<string>();
-
-    const addFeatures = async (q: string) => {
-      const result = await e.query(q);
-      const it = result.iter({feature: STR});
-      for (; it.valid(); it.next()) {
-        features.add(it.feature);
-      }
-    };
-
-    await addFeatures(`
-      select distinct 'atom.' || s.name as feature
-      from track t join slice s on t.id = s.track_id
-      where t.name = 'Statsd Atoms'`);
-
-    await addFeatures(`
-      select distinct
-        case when name like '%wlan%' then 'net.wifi'
-            when name like '%rmnet%' then 'net.modem'
-            else 'net.other'
-        end as feature
-      from track
-      where name like '%Transmitted' or name like '%Received'`);
-
-    await addFeatures(`
-      select distinct 'track.' || lower(name) as feature
-      from track where name in ('RIL', 'suspend_backoff') or name like 'battery_stats.%'`);
-
-    await addFeatures(`
-      select distinct 'track.battery_stats.*' as feature
-      from track where name like 'battery_stats.%'`);
-
-    try {
-      await e.query(
-        `INCLUDE PERFETTO MODULE
-              google3.wireless.android.telemetry.trace_extractor.modules.atom_counters_slices`,
-      );
-      features.add('google3');
-    } catch (e) {}
-
-    return features;
-  }
-
-  async addDayExplorerCommand(
-    ctx: Trace,
-    features: Set<string>,
-  ): Promise<void> {
-    if (features.has('google3')) {
-      ctx.commands.registerCommand({
-        id: 'com.android.DayExplorerBlamesByCategory',
-        name: 'Add tracks: Day Explorer',
-        callback: async () => {
-          const limitStr = await ctx.omnibox.prompt(
-            'Maximum results per group',
-          );
-          const limit = Number(limitStr);
-          if (!isFinite(limit) || limit <= 0) {
-            alert('Positive number required');
-            return;
-          }
-          await this.addDayExplorerBehaviors(ctx, 'Day explorer');
-          await this.addDayExplorerCounters(ctx, 'Day explorer', limit);
-        },
-      });
-    }
-  }
-
   async onTraceLoad(ctx: Trace): Promise<void> {
-    const features: Set<string> = await this.findFeatures(ctx.engine);
-    const containedTraces = (ctx.openerPluginArgs?.containedTraces ??
-      []) as ContainedTrace[];
+    const support = this.support(ctx);
+    const features = await support.features(ctx.engine);
 
     await ctx.engine.query(PACKAGE_LOOKUP);
-    await this.addNetworkSummary(ctx, features);
-    await this.addBluetooth(ctx, features);
-    await this.addModemRil(ctx, features);
-    await this.addKernelWakelocks(ctx);
-    await this.addKernelWakelocksStatsd(ctx, features);
-    await this.addWakeups(ctx, features);
-    await this.addDeviceState(ctx, features);
-    await this.addHighCpu(ctx, features);
-    await this.addContainedTraces(ctx, containedTraces);
-    ctx.selection.registerAreaSelectionTab(
-      this.createDayExplorerFlameGraphPanel(ctx),
-    );
+    await this.addNetworkSummary(ctx, support, features);
+    await this.addKernelWakelocks(ctx, support);
+    await this.addKernelWakelocksStatsd(ctx, support, features);
+    await this.addWakeups(ctx, support, features);
+    await this.addDeviceState(ctx, support, features);
+    await this.addHighCpu(ctx, support, features);
 
     if (features.has('google3')) {
-      await this.addAtomCounters(ctx);
-      await this.addAtomSlices(ctx);
-      await this.addModemTeaData(ctx);
-      await this.addDayExplorerCommand(ctx, features);
+      await this.addAtomCounters(ctx, support);
+      await this.addAtomSlices(ctx, support);
+      await this.addModemTeaData(ctx, support);
     }
   }
 }
