+++ ui/src/plugins/com.android.RilModem/OWNERS
@@ -1,0 +1,2 @@
+simonmacm@google.com
+rzuklie@google.com
